Added Changes

+++ Closure2/Closure2/Models/Branch.cs
@@ -10,8 +10,8 @@ namespace Closure2.Models
         public int ID { set; get; }
         public string Name { set; get; }
         public string Street { set; get; }
-        public int latitude { set; get; }
-        public int longitude { set; get; }
+        public float latitude { set; get; }
+        public float longitude { set; get; }
         public virtual ICollection<ProductModels> Products { get; set; }
     }
 }
